backend/requirements.txt

flask==2.0.1
flask-cors==3.0.10
transformers>=4.41.1
tokenizers>=0.19.0,<0.20
pillow==9.5.0
tabulate==0.8.10
pandas==2.0.3
requests==2.31.0
torch==2.1.0
numpy>=1.26.0
werkzeug==2.0.1
